+++ interpreter/emulator/pong/__init__.py
@@ -1,0 +1,1 @@
+"""Batocera/Knulli-oriented Pong launcher assets."""